+++ libbitcoinkernel-sys/bitcoin/src/txdb.h
@@ -52,8 +52,7 @@ public:
     uint256 GetBestBlock() const override;
     std::vector<uint256> GetHeadBlocks() const override;
     void BatchWrite(CoinsViewCacheCursor& cursor, const uint256& block_hash) override;
-    //! Get a cursor to iterate over the whole state.
-    std::unique_ptr<CCoinsViewCursor> Cursor() const;
+    std::unique_ptr<CCoinsViewCursor> Cursor() const override;
 
     //! Whether an unsupported database format is used.
     bool NeedsUpgrade();