+++ src/state/mod.rs
@@ -1,9 +1,12 @@
 pub mod chain;
 pub mod chainstate;
+pub mod coins;
 pub mod context;
 
 pub use chain::{Chain, ChainIterator};
 pub use chainstate::{
     ChainstateManager, ChainstateManagerBuilder, ProcessBlockHeaderResult, ProcessBlockResult,
+    ValidateBlockResult,
 };
+pub use coins::FetchCoinCallback;
 pub use context::{ChainParams, ChainType, Context, ContextBuilder};